refactor: split keyboard shortcut config
